Refactoring Part 4

See merge request lgln/power.ni/power-frontend!15

+++ power/src/app/fragebogen/dashboard/storage.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '@env/environment';
-import {options} from 'knockout';
+import { Observable } from 'rxjs';
 
 /**
  * StorageService handles api requests and data storage
@@ -10,9 +10,11 @@ import {options} from 'knockout';
   providedIn: 'root'
 })
 export class StorageService {
-  public formsList: any = [];
-  public tagList: any = [];
-  public serviceList: any = [{value: 'AKS', name: 'Automatische Kaufpreissammlung'}];
+  public formsList = [];
+  public tagList = [];
+  public serviceList = [
+    {value: 'AKS', name: 'Automatische Kaufpreissammlung'}
+  ];
 
   constructor(private httpClient: HttpClient) {
   }
@@ -28,16 +30,22 @@ export class StorageService {
   /**
    * Loads list of forms
    */
-  public loadFormsList() {
+  public loadFormsList(): Observable<Object> {
     // Load data from server
-    const url = environment.formAPI + 'intern/forms?fields=access,access-minutes,created,id,owners,readers,status,tags,title';
+    const url = environment.formAPI
+                + 'intern/forms?fields=access,access-minutes,created,id,owners,readers,status,tags,title';
     return this.httpClient.get(url);
   }
 
   /**
    * Loads a form by id
    */
-  public loadForm(id: string) {
+  public loadForm(id: string): Observable<Object> {
+    // check data
+    if (!id) {
+      throw new Error('id is required');
+    }
+
     // load form from server
     const url = environment.formAPI + 'intern/forms/' + encodeURIComponent(id);
     return this.httpClient.get(url);
@@ -46,7 +54,7 @@ export class StorageService {
   /**
    * Load tags
    */
-  public loadTags() {
+  public loadTags(): Observable<Object> {
     // Load tags from server
     const url = environment.formAPI + 'intern/tags';
     return this.httpClient.get(url);
@@ -57,9 +65,17 @@ export class StorageService {
    * @param data SurveyJS
    * @param tags Tag list
    */
-  public createForm(data: any, tags?: string) {
+  public createForm(data: any, tags?: string): Observable<Object> {
+    // check data
+    if (!data) {
+      throw new Error('data is required');
+    }
+
     // Uploads form
-    const url = environment.formAPI + 'intern/forms' + (tags ? '?tags=' + tags : '');
+    let url = environment.formAPI + 'intern/forms';
+    if (tags) {
+      url += '?tags=' + encodeURIComponent(tags);
+    }
     return this.httpClient.post(url, data, {
       headers: {
         'Content-Type': 'application/json'
@@ -71,7 +87,12 @@ export class StorageService {
    * Deletes form
    * @param id Form id
    */
-  public deleteForm(id: string) {
+  public deleteForm(id: string): Observable<Object> {
+    // check data
+    if (!id) {
+      throw new Error('id is required');
+    }
+
     // Delete form
     const url = environment.formAPI + 'intern/forms/' + encodeURIComponent(id);
     return this.httpClient.delete(url);
